+++ SistemaInventario/src/main/java/EntornosProgramacion/SistemaInventario/repository/ArqueoCajaRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface ArqueoCajaRepository extends JpaRepository<ArqueoCaja, Integer> {
+public interface ArqueoCajaRepository extends JpaRepository<ArqueoCaja, Long> {
 
     Optional<ArqueoCaja> findByUsuarioIdAndAbiertoTrue(Long usuarioId);
